requirements.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
onnx==1.15.0
onnxruntime==1.16.3
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.2
Pillow==10.2.0
tqdm==4.66.1
pyyaml==6.0.1
tensorboard==2.15.2
thop==0.1.1-2209072238  # for FLOPs estimation
roboflow==1.1.9          # for dataset download (optional)
scipy==1.12.0
pytest==7.4.4
jupyter==1.0.0

# Optional for pruning visualization
torchinfo==1.8.0
